requirements.txt
ccxt>=4.0.0
pandas>=2.2.0
pyarrow>=15.0.0
python-dotenv>=1.0.1